+++ relay-server/src/actors/processor.rs
@@ -2632,15 +2632,23 @@ impl EnvelopeProcessorService {
                 let mut timestamp = item.timestamp().unwrap_or(received_timestamp);
                 clock_drift_processor.process_timestamp(&mut timestamp);
 
-                let metrics =
-                    Metric::parse_all(&payload, timestamp).filter_map(|result| result.ok());
+                let mut buckets = Vec::new();
+                for bucket_result in Bucket::parse_all(&payload, timestamp) {
+                    match bucket_result {
+                        Ok(bucket) => buckets.push(bucket),
+                        Err(error) => relay_log::debug!(
+                            error = &error as &dyn Error,
+                            "failed to parse metric bucket from statsd format",
+                        ),
+                    }
+                }
 
-                relay_log::trace!("inserting metrics into project cache");
+                relay_log::trace!("inserting metric buckets into project cache");
                 self.inner
                     .project_cache
-                    .send(InsertMetrics::new(public_key, metrics));
+                    .send(MergeBuckets::new(public_key, buckets));
             } else if item.ty() == &ItemType::MetricBuckets {
-                match Bucket::parse_all(&payload) {
+                match serde_json::from_slice::<Vec<Bucket>>(&payload) {
                     Ok(mut buckets) => {
                         for bucket in &mut buckets {
                             clock_drift_processor.process_timestamp(&mut bucket.timestamp);